requirements.txt
# PockNet - Protein Binding Site Prediction Requirements
# Complete pip requirements based on fully tested pocknet_env
# This file contains all Python packages needed for:
# - PyTorch Lightning training with Hydra configuration
# - P2Rank protein feature extraction pipeline  
# - ESM protein language model embeddings
# - Jupyter notebook analysis and visualization
# - WandB experiment tracking

# --------- Core PyTorch ecosystem --------- #
torch==2.6.0
torchvision==0.21.0
torchmetrics==1.6.1
torchinfo==1.8.0
triton==3.2.0

# --------- NVIDIA CUDA dependencies --------- #
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127

# --------- Lightning ecosystem --------- #
pytorch-lightning==2.5.0.post0
lightning==2.5.0.post0
lightning-utilities==0.12.0

# --------- TabNet and attention mechanisms --------- #
pytorch-tabnet==4.1.0
entmax==1.3

# --------- Configuration management --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0
omegaconf==2.3.0

# --------- ESM protein language model --------- #
fair-esm==2.0.0

# --------- Scientific computing --------- #
numpy>=1.26.4
scipy>=1.15.2
scikit-learn>=1.6.1
pandas==2.3.1

# --------- Data processing --------- #
h5py>=3.12.1
pyarrow==19.0.1
fastparquet==2024.11.0

# --------- Visualization --------- #
matplotlib>=3.9.2
seaborn>=0.13.2

# --------- Jupyter ecosystem --------- #
jupyter>=1.0.0
jupyterlab>=4.3.4
ipykernel>=6.29.5
ipython>=8.31.0
notebook>=7.3.2
ipywidgets==8.1.5
jupyterlab-widgets==3.0.13
widgetsnbextension==4.0.13

# --------- Protein analysis tools --------- #
biopython==1.85

# --------- Protein visualization --------- #
py3dmol==2.4.2
nglview==3.1.4

# --------- WEKA wrapper for P2Rank --------- #
python-weka-wrapper3==0.3.2
jpype1==1.5.2
liac-arff==2.5.0

# --------- Machine learning utilities --------- #
optuna==2.10.1
imbalanced-learn==0.13.0

# --------- Experiment tracking --------- #
wandb==0.19.4
tensorboard==2.19.0
tensorboard-data-server==0.7.2

# --------- Development tools --------- #
pre-commit==4.1.0
pytest==8.3.4
rich==13.9.4
rootutils==1.0.7

# --------- Additional utilities --------- #
python-dotenv==1.0.1
click==8.1.8
pydantic==2.10.5
pydantic-core==2.27.2
gitpython==3.1.44
markdown==3.8.2
markdown-it-py==3.0.0
mdurl==0.1.2
tqdm>=4.67.1
psutil>=6.1.1

# --------- Web and HTTP utilities --------- #
requests>=2.32.3
urllib3>=2.3.0
aiohttp==3.11.11
httpx>=0.28.1

# --------- Configuration and data formats --------- #
pyyaml>=6.0.2
toml>=0.10.2
tomli>=2.2.1

# --------- Development dependencies --------- #
setuptools>=75.8.0
wheel>=0.45.1
packaging>=24.2

# --------- Additional scientific packages --------- #
sympy==1.13.1
networkx==3.4.2
mpmath==1.3.0
